Factored out test utility functions to avoid importing test files from other tests.
